Rename files to lowercase for consistency
